+++ ui/src/main/java/com/gallery/ui/result/Types.kt
@@ -4,9 +4,3 @@ enum class FinderType {
     POPUP,
     BOTTOM
 }
-
-enum class CropType {
-    @Deprecated("annoying version support", replaceWith = ReplaceWith("CROPPER"))
-    UCROP,
-    CROPPER
-}